+++ runtime/giphy.js
@@ -1,3 +1,6 @@
+var qs = require("querystring");
+var Logger = require("./logger.js").Logger;
+
 exports.get_gif = function(tags, func) {
   //limit=1 will only return 1 gif
   var params = {
@@ -16,7 +19,7 @@ exports.get_gif = function(tags, func) {
   var request = require("request");
   //Logger.log("debug", query)
 
-  request(config.url + "?" + query, function(error, response, body) {
+  request("http://api.giphy.com/v1/gifs/search" + "?" + query, function(error, response, body) {
     //Logger.log("debug", arguments)
     if (error || response.statusCode !== 200) {
       Logger.log("error", "giphy: Got error: " + body);