+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.0-beta.3
+Fixed `giphy.js` not having requires.  
+Fixed `suffix` not behaving accordingly.    
+Added `++fortunecow`, `++randomcat`, `++rule34` and `++leetspeak`.   
+Added `mashape_key` to `config.json`.   
+Removed NSFW flags from commands that did not need them, as this will cause problems with `++help`.  
+Removed `deletion.js`, as this is needlessly split.   
+
 ## 2.0.0-beta.2
 Updated `versionchecker.js` to check for beta updates.   
 Moved to double prefix activation instead of single prefix activation as requested by Discord API, we recommend using `++`.